Version 1.3-beta 1
Changes:
- Added functions to locally check the current version with the new one (information is taken from the main PE file). The code for the activation is blocked for the time being. other functions have not yet been made
- All functions from the main code are moved to another file to offload the main code
- Made changes for future beta versions
