+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gru"
-version = "1.2.0"
+version = "1.3.0"
 edition = "2021"
 authors = ["Zalexanninev15 <blue.shark@disroot.org>"]
 license = "MIT License"
@@ -18,6 +18,8 @@ winapi = { version = "0.3.9", features = [
   "impl-default",
 ] }
 arguments = "0.6.2"
+curl = "0.4.44"
+pelite = "0.9.1"
 
 [target.x86_64-pc-windows-msvc]
 rustflags = ["-C", "target-feature=+crt-static"]
